unit/lib/init: Use short int for the %d directive
scan-build under fedora40 reports the following:

subsystem_ut.c:69:76: error: ‘%d’ directive output may be truncated
writing between 1 and 10 bytes into a region of size 7
[-Werror=format-truncation=]

The subsystem_name at minimum requires 10 bytes and there's a space
for 6 more. However, %d under int  may output up to 10 bytes.
Changing it to short int drops the max output to 5 bytes which fits
into the existing buffer nicely.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24647
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ test/unit/lib/init/subsystem.c/subsystem_ut.c
@@ -57,7 +57,7 @@ static void
 subsystem_sort_test_depends_on_single(void)
 {
 	struct spdk_subsystem *subsystem;
-	int i;
+	short int i;
 	char subsystem_name[16];
 
 	global_rc = -1;
